WIP: Fix verifying all the alloc requests from a multi-create
Previously we were always testing the first allocation request from a list,
which was then testing the same allocation request in case of a multi-create.

If the capacity wasn't enough from that host for accepting all the instances,
then we were returning a NoValidHost instead of verifying the other resources.

Change-Id: Idf2478b0b7b45053bd9828ff0a046c827e766837
Closes-Bug: #1874664